[DEV-1008] Create Security YAML to condense the pipelines (#116) (#117)
* [DEV-1008] Create Security YAML to condense the pipelines (#116)

* Create security.yml

* Update security.yml

* Update security.yml

* Create email-report.yml

* Update security.yml

* Delete .github/workflows/security-scan.yml

* Update security.yml

* Update security.yml

* updated to ignore local properties file.

* Create new CCD endpoints for CompTox Chemicals Dashboard #118 (#119)

* API for biomonitoring data #87 (#120)

* Update CCDBiomonitoring.java (#122)

this changed did not stage before commit

* Create CCD projection for biomonitoring #121 (#123)

* Enhance README.md with detailed project overview, technologies, configuration, and getting started instructions

* Update data source for Qsur resource #124 (#125)

* Update data source for Qsur resource #124

* typo in biomonitoring source name

* Update CCDBiomonitoring.java (#126)

typo in data source name. i already updated the schema name in dbeaver

* Create endpoint for CCDProductionVolume Resource #127 (#128)

* Create endpoint for CCDProductionVolume Resource #127

* Updates from editor

* Update CCDBiomonitoringRepository.java (#129)

* Create ccd projections for GenExpoPredicitons and (#131)

DemoExpoPredictions #130

* Update README.md for clarity and improved project documentation

* Add MIT License file to the project

* Add CONTRIBUTING.md to provide guidelines for project contributions

* Add Code of Conduct to outline community expectations and standards

* Add initial code.json file for CTX Exposure APIs with metadata and licensing information

* Update schema name for CCDProductionVolume table to 'prodvol'

* Implement 'paging' for MMDB by-medium endpoints #110 (#132)

* Implement 'paging' for MMDB by-medium endpoints #110

* Update MMDBAggregateRepository.java

* Update MMDBSingleSampleRepository.java

* Update MMDBService.java (#133)

* Update MMDBService.java

updating variables on pages

* Update MMDBPage.java

* Update Swagger Documentation #134 (#135)

* Update Exposure API version to 2.0.0 (#137)

* Bridirks patch 6 (#138)

* Update Dockerfile

* Update manifest-prod.yml

* Update manifest-stg.yml

* Update pom.xml

---------

+++ src/main/java/gov/epa/ccte/api/exposure/domain/MMDBAggregate.java
@@ -7,7 +7,6 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Getter;
 import lombok.Setter;
@@ -21,7 +20,7 @@ public class MMDBAggregate {
 	@Id
 	@Column(name = "id")
 	@Schema(description ="Unique database ID for this individual table record")
-	private Integer id;
+	private Long id;
 
 	@Size(max = 120)
 	@Column(name = "full_source_name", length = 120)
